+++ libcms/apps/participant_menu/administration/views.py
@@ -8,18 +8,23 @@ from django.contrib.auth.decorators import login_required
 from django.utils.translation import get_language
 
 from common.pagination import get_page
-from ..models import Menu, MenuTitle, MenuItem, MenuItemTitle
-from forms import MenuForm,MenuTitleForm,  MenuItemForm, MenuItemTitleForm
 
-from participant_site import decorators
+from participants import decorators, org_utils
+
+from ..models import Menu, MenuTitle, MenuItem, MenuItemTitle
+from forms import MenuForm, MenuTitleForm, MenuItemForm, MenuItemTitleForm
 
 
 @login_required
 @transaction.atomic()
-@decorators.must_be_manager
-def index(request, library_code, library):
+@decorators.must_be_org_user
+def index(request, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     if not request.user.has_module_perms('participant_menu'):
-        return HttpResponseForbidden()
+        return HttpResponseForbidden(u'У вас нет доступа к разделу')
     try:
         Menu.objects.get(slug='main_menu', library=library)
     except Menu.DoesNotExist:
@@ -33,13 +38,14 @@ def index(request, library_code, library):
     return redirect('participant_menu:administration:menu_list', library_code=library_code)
 
 
-
-
-
 @login_required
-#@permission_required_or_403('menu.add_menu')
-@decorators.must_be_manager
-def menu_list(request, library_code, library):
+# @permission_required_or_403('menu.add_menu')
+@decorators.must_be_org_user
+def menu_list(request, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     if not request.user.has_module_perms('participant_menu'):
         return HttpResponseForbidden()
     menus_page = get_page(request, Menu.objects.filter(library=library))
@@ -47,7 +53,7 @@ def menu_list(request, library_code, library):
 
     menus_dict = {}
     for menu in menus_page.object_list:
-        menus_dict[menu.id] = {'menu':menu}
+        menus_dict[menu.id] = {'menu': menu}
 
     for menu_title in menu_titles:
         menus_dict[menu_title.menu_id]['menu'].menu_title = menu_title
@@ -58,16 +64,17 @@ def menu_list(request, library_code, library):
         'library': library,
         'menus': menus,
         'menus_page': menus_page,
-        })
-
-
+    })
 
 
 @login_required
 @permission_required_or_403('participant_menu.add_menu')
 @transaction.atomic()
-@decorators.must_be_manager
-def create_menu(request, library_code, library):
+@decorators.must_be_org_user
+def create_menu(request, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
 
     if request.method == 'POST':
 
@@ -76,7 +83,7 @@ def create_menu(request, library_code, library):
 
         for lang in settings.LANGUAGES:
             menu_title_forms.append({
-                'form':MenuTitleForm(
+                'form': MenuTitleForm(
                     request.POST,
                     prefix="menu_title_" + lang[0]
                 ),
@@ -109,13 +116,13 @@ def create_menu(request, library_code, library):
         menu_title_forms = []
         for lang in settings.LANGUAGES:
             menu_title_forms.append({
-                'form':MenuTitleForm(
+                'form': MenuTitleForm(
                     initial={
-                        'lang':lang[0]
+                        'lang': lang[0]
                     },
                     prefix="menu_title_" + lang[0]
                 ),
-                'lang':lang
+                'lang': lang
             })
         menu_form = MenuForm(prefix='menu_form')
 
@@ -126,13 +133,14 @@ def create_menu(request, library_code, library):
     })
 
 
-
-
 @login_required
 @permission_required_or_403('participant_menu.change_menu')
 @transaction.atomic
-@decorators.must_be_manager
-def edit_menu(request, id, library_code, library):
+@decorators.must_be_org_user
+def edit_menu(request, id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
 
     menu = get_object_or_404(Menu, id=id)
     menu_item_titles = MenuTitle.objects.filter(menu=menu)
@@ -145,20 +153,17 @@ def edit_menu(request, id, library_code, library):
 
         menu_title_forms = []
 
-
-
         if menu_form.is_valid():
             menu = menu_form.save(commit=False)
 
             for lang in settings.LANGUAGES:
                 menu_title_forms.append({
-                    'form':MenuTitleForm(
+                    'form': MenuTitleForm(
                         request.POST,
                         prefix="menu_title_" + lang[0]
                     ),
                     'lang': lang
                 })
-
 
             valid = False
             for menu_title_form in menu_title_forms:
@@ -168,10 +173,9 @@ def edit_menu(request, id, library_code, library):
             if valid:
                 menu.save()
 
-
                 for menu_title_form in menu_title_forms:
                     lang = menu_title_form['form'].cleaned_data['lang']
-                    if lang in  menu_item_titles_langs:
+                    if lang in menu_item_titles_langs:
                         if menu_item_titles_langs[lang].title != menu_title_form['form'].cleaned_data['title']:
                             menu_item_titles_langs[lang].title = menu_title_form['form'].cleaned_data['title']
                             menu_item_titles_langs[lang].save()
@@ -192,11 +196,11 @@ def edit_menu(request, id, library_code, library):
                 if menu_item_title.lang == lang[0]:
                     menus_title_langs.append(lang[0])
                     menu_title_forms.append({
-                        'form':MenuTitleForm(
+                        'form': MenuTitleForm(
                             initial={
                                 'lang': menu_item_title.lang,
                                 'title': menu_item_title.title,
-                                },
+                            },
                             prefix="menu_title_" + menu_item_title.lang
                         ),
                         'lang': menu_item_title.lang
@@ -204,22 +208,21 @@ def edit_menu(request, id, library_code, library):
         new_langs = []
         if len(settings.LANGUAGES) != len(menus_title_langs):
             for lang in settings.LANGUAGES:
-                if lang[0] not in  menus_title_langs:
+                if lang[0] not in menus_title_langs:
                     new_langs.append(lang[0])
 
         for lang in new_langs:
             menu_title_forms.append({
-                'form':MenuTitleForm(
+                'form': MenuTitleForm(
                     initial={
-                        'lang':lang
+                        'lang': lang
                     },
                     prefix="menu_title_" + lang
                 ),
-                'lang':lang
+                'lang': lang
             })
 
         menu_form = MenuForm(prefix='menu_form', instance=menu)
-
 
     return render(request, 'participant_menu/administration/edit_menu.html', {
         'library': library,
@@ -231,22 +234,28 @@ def edit_menu(request, id, library_code, library):
 @login_required
 @permission_required_or_403('participant_menu.delete_menu')
 @transaction.atomic
-@decorators.must_be_manager
-def delete_menu(request, id, library_code, library):
+@decorators.must_be_org_user
+def delete_menu(request, id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     menu = get_object_or_404(Menu, id=id)
     menu.delete()
     return redirect('participant_menu:administration:menus_list', library_code=library_code)
 
 
-
-
 @login_required
-#@permission_required_or_403('menu.create_menu')
-@decorators.must_be_manager
-def item_list(request, menu_id, library_code, library):
+# @permission_required_or_403('menu.create_menu')
+@decorators.must_be_org_user
+def item_list(request, menu_id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     menu = get_object_or_404(Menu, id=menu_id)
     nodes = list(menu.root_item.get_descendants())
-    lang=get_language()[:2]
+    lang = get_language()[:2]
     item_titles = MenuItemTitle.objects.filter(item__in=nodes, lang=lang)
     nd = {}
     for node in nodes:
@@ -265,8 +274,12 @@ def item_list(request, menu_id, library_code, library):
 @login_required
 @permission_required_or_403('participant_menu.add_menuitem')
 @transaction.atomic
-@decorators.must_be_manager
-def create_item(request, menu_id, library_code, library, parent=None):
+@decorators.must_be_org_user
+def create_item(request, menu_id, library_code, parent=None, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     menu = get_object_or_404(Menu, id=menu_id)
 
     if not parent:
@@ -275,13 +288,13 @@ def create_item(request, menu_id, library_code, library, parent=None):
         parent = get_object_or_404(MenuItem, id=parent)
 
     if request.method == 'POST':
-        item_form = MenuItemForm(request.POST,prefix='item_form')
+        item_form = MenuItemForm(request.POST, prefix='item_form')
 
         menu_item_title_forms = []
         for lang in settings.LANGUAGES:
             menu_item_title_forms.append({
-                'form':MenuItemTitleForm(request.POST,prefix='menu_item_title_' + lang[0]),
-                'lang':lang[0]
+                'form': MenuItemTitleForm(request.POST, prefix='menu_item_title_' + lang[0]),
+                'lang': lang[0]
             })
 
         if item_form.is_valid():
@@ -309,23 +322,27 @@ def create_item(request, menu_id, library_code, library, parent=None):
         menu_item_title_forms = []
         for lang in settings.LANGUAGES:
             menu_item_title_forms.append({
-                'form':MenuItemTitleForm(initial={'url': u'/'+lang[0]+u'/#'},prefix='menu_item_title_' + lang[0]),
-                'lang':lang[0]
+                'form': MenuItemTitleForm(initial={'url': u'/' + lang[0] + u'/#'}, prefix='menu_item_title_' + lang[0]),
+                'lang': lang[0]
             })
 
     return render(request, 'participant_menu/administration/create_item.html', {
         'library': library,
         'item_form': item_form,
         'menu_item_title_forms': menu_item_title_forms,
-        'menu':menu
+        'menu': menu
     })
 
 
 @login_required
 @permission_required_or_403('participant_menu.change_menuitem')
 @transaction.atomic()
-@decorators.must_be_manager
-def item_edit(request, id, library_code, library, menu_id=None):
+@decorators.must_be_org_user
+def item_edit(request, id, library_code, menu_id=None, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     menu = get_object_or_404(Menu, id=menu_id)
     item = get_object_or_404(MenuItem, id=id)
     item_titles = MenuItemTitle.objects.filter(item=item)
@@ -334,10 +351,8 @@ def item_edit(request, id, library_code, library, menu_id=None):
     for lang in settings.LANGUAGES:
         item_titles_langs[lang] = None
 
-
     for item_title in item_titles:
         item_titles_langs[item_title.lang] = item_title
-
 
     if request.method == 'POST':
         item_form = MenuItemForm(request.POST, prefix='item_form', instance=item)
@@ -347,13 +362,14 @@ def item_edit(request, id, library_code, library, menu_id=None):
                 lang = lang[0]
                 if lang in item_titles_langs:
                     menu_item_title_forms.append({
-                        'form':MenuItemTitleForm(request.POST,prefix='menu_item_title_' + lang, instance=item_titles_langs[lang]),
-                        'lang':lang
+                        'form': MenuItemTitleForm(request.POST, prefix='menu_item_title_' + lang,
+                                                  instance=item_titles_langs[lang]),
+                        'lang': lang
                     })
                 else:
                     menu_item_title_forms.append({
-                        'form':MenuItemTitleForm(request.POST,prefix='menu_item_title_' + lang),
-                        'lang':lang
+                        'form': MenuItemTitleForm(request.POST, prefix='menu_item_title_' + lang),
+                        'lang': lang
                     })
 
         valid = False
@@ -361,7 +377,6 @@ def item_edit(request, id, library_code, library, menu_id=None):
             valid = menu_item_title_form['form'].is_valid()
             if not valid:
                 break
-
 
         if not item_form.is_valid():
             valid = False
@@ -390,13 +405,13 @@ def item_edit(request, id, library_code, library, menu_id=None):
             lang = lang[0]
             if lang in item_titles_langs:
                 menu_item_title_forms.append({
-                    'form':MenuItemTitleForm(prefix='menu_item_title_' + lang, instance=item_titles_langs[lang]),
-                    'lang':lang
+                    'form': MenuItemTitleForm(prefix='menu_item_title_' + lang, instance=item_titles_langs[lang]),
+                    'lang': lang
                 })
             else:
                 menu_item_title_forms.append({
-                    'form':MenuItemTitleForm(prefix='menu_item_title_' + lang),
-                    'lang':lang
+                    'form': MenuItemTitleForm(prefix='menu_item_title_' + lang),
+                    'lang': lang
                 })
 
     return render(request, 'participant_menu/administration/edit_item.html', {
@@ -407,29 +422,44 @@ def item_edit(request, id, library_code, library, menu_id=None):
         'menu': menu
     })
 
+
 @login_required
 @permission_required_or_403('participant_menu.delete_menuitem')
 @transaction.atomic
-@decorators.must_be_manager
-def item_delete(request, menu_id, id, library_code, library):
+@decorators.must_be_org_user
+def item_delete(request, menu_id, id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     item = get_object_or_404(MenuItem, id=id)
     item.delete()
     return redirect('participant_menu:administration:item_list', menu_id=menu_id, library_code=library_code)
 
-@login_required
-@permission_required_or_403('participant_menu.change_menu')
-@transaction.atomic
-@decorators.must_be_manager
-def item_up(request, menu_id, id, library_code, library):
-    item = get_object_or_404(MenuItem, id=id)
-    item.up()
-    return redirect('participant_menu:administration:item_list', menu_id=menu_id, library_code=library_code)
 
 @login_required
 @permission_required_or_403('participant_menu.change_menu')
 @transaction.atomic
-@decorators.must_be_manager
-def item_down(request, menu_id, id, library_code, library):
+@decorators.must_be_org_user
+def item_up(request, menu_id, id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
+    item = get_object_or_404(MenuItem, id=id)
+    item.up()
+    return redirect('participant_menu:administration:item_list', menu_id=menu_id, library_code=library_code)
+
+
+@login_required
+@permission_required_or_403('participant_menu.change_menu')
+@transaction.atomic
+@decorators.must_be_org_user
+def item_down(request, menu_id, id, library_code, managed_libraries=[]):
+    library = org_utils.get_library(library_code, managed_libraries)
+    if not library:
+        return HttpResponseForbidden(u'Вы должны быть сотрудником этой организации')
+
     item = get_object_or_404(MenuItem, id=id)
     item.down()
     return redirect('participant_menu:administration:item_list', menu_id=menu_id, library_code=library_code)
